Extracted test functionality to separate folder
