+++ src/app/models/AudioEvent.ts
@@ -1,7 +1,6 @@
 import { Injector } from "@angular/core";
 import { AUDIO_RECORDING } from "@baw-api/ServiceTokens";
 import { DateTimeTimezone, Id } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, Deleter, HasOne, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -59,26 +58,6 @@ export class AudioEvent extends AbstractModel implements IAudioEvent {
   public deleter?: User;
   @HasOne<AudioEvent>(AUDIO_RECORDING, "audioRecordingId")
   public audioRecording?: AudioRecording;
-
-  public static generate(id?: Id): IAudioEvent {
-    const hertzRange = [22050, 8000, 44100, 48000];
-
-    return {
-      id: modelData.id(id),
-      audioRecordingId: modelData.id(),
-      startTimeSeconds: modelData.seconds(),
-      endTimeSeconds: modelData.seconds(),
-      lowFrequencyHertz: modelData.random.arrayElement(hertzRange),
-      highFrequencyHertz: modelData.random.arrayElement(hertzRange),
-      isReference: modelData.boolean(),
-      creatorId: modelData.id(),
-      updaterId: modelData.id(),
-      deleterId: modelData.id(),
-      createdAt: modelData.timestamp(),
-      updatedAt: modelData.timestamp(),
-      deletedAt: modelData.timestamp(),
-    };
-  }
 
   constructor(audioEvent: IAudioEvent, injector?: Injector) {
     super(audioEvent, injector);
