+++ src/app/models/Bookmark.ts
@@ -6,7 +6,6 @@ import {
   Id,
   Param,
 } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, HasOne, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -57,21 +56,6 @@ export class Bookmark extends AbstractModel implements IBookmark {
   public updater?: User;
   @HasOne<Bookmark>(AUDIO_RECORDING, "audioRecordingId")
   public audioRecording?: AudioRecording;
-
-  public static generate(id?: Id): IBookmark {
-    return {
-      id: modelData.id(id),
-      audioRecordingId: modelData.id(),
-      offsetSeconds: modelData.seconds(),
-      name: modelData.param(),
-      creatorId: modelData.id(),
-      updaterId: modelData.id(),
-      createdAt: modelData.timestamp(),
-      updatedAt: modelData.timestamp(),
-      description: modelData.description(),
-      category: "<< application >>", // TODO Replace with list of possibilities
-    };
-  }
 
   constructor(bookmark: IBookmark, injector?: Injector) {
     super(bookmark, injector);
