+++ src/app/models/Dataset.ts
@@ -5,7 +5,6 @@ import {
   Id,
   Param,
 } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -41,18 +40,6 @@ export class Dataset extends AbstractModel implements IDataset {
   public creator: User;
   @Updater<Dataset>()
   public updater?: User;
-
-  public static generate(id?: Id): IDataset {
-    return {
-      id: modelData.id(id),
-      name: modelData.param(),
-      description: modelData.description(),
-      creatorId: modelData.id(),
-      updaterId: modelData.id(),
-      createdAt: modelData.timestamp(),
-      updatedAt: modelData.timestamp(),
-    };
-  }
 
   constructor(dataset: IDataset, injector?: Injector) {
     super(dataset, injector);
