+++ src/app/models/DatasetItem.ts
@@ -1,7 +1,6 @@
 import { Injector } from "@angular/core";
 import { AUDIO_RECORDING, DATASET } from "@baw-api/ServiceTokens";
 import { DateTimeTimezone, Id } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, HasOne } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -45,19 +44,6 @@ export class DatasetItem extends AbstractModel implements IDatasetItem {
   public dataset?: Dataset;
   @HasOne<DatasetItem>(AUDIO_RECORDING, "audioRecordingId")
   public audioRecording?: AudioRecording;
-
-  public static generate(id?: Id): IDatasetItem {
-    return {
-      id: modelData.id(id),
-      datasetId: modelData.id(),
-      audioRecordingId: modelData.id(),
-      creatorId: modelData.id(),
-      createdAt: modelData.timestamp(),
-      startTimeSeconds: modelData.seconds(),
-      endTimeSeconds: modelData.seconds(),
-      order: modelData.random.number(100),
-    };
-  }
 
   constructor(datasetItem: IDatasetItem, injector?: Injector) {
     super(datasetItem, injector);
