+++ src/app/models/ProgressEvent.ts
@@ -1,7 +1,6 @@
 import { Injector } from "@angular/core";
 import { DATASET_ITEM } from "@baw-api/ServiceTokens";
 import { DateTimeTimezone, Id } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, HasOne } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -33,16 +32,6 @@ export class ProgressEvent extends AbstractModel implements IProgressEvent {
   public creator?: User;
   @HasOne<ProgressEvent>(DATASET_ITEM, "datasetItemId")
   public datasetItem?: DatasetItem;
-
-  public static generate(id?: Id): IProgressEvent {
-    return {
-      id: modelData.id(id),
-      creatorId: modelData.id(),
-      datasetItemId: modelData.id(),
-      activity: "viewed", // TODO Replace with list of possibilities
-      createdAt: modelData.timestamp(),
-    };
-  }
 
   constructor(progressEvent: IProgressEvent, injector?: Injector) {
     super(progressEvent, injector);
