+++ src/app/models/Question.ts
@@ -1,6 +1,5 @@
 import { Injector } from "@angular/core";
 import { DateTimeTimezone, Id } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, Updater } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -36,18 +35,6 @@ export class Question extends AbstractModel implements IQuestion {
   public creator?: User;
   @Updater<Question>()
   public updater?: User;
-
-  public static generate(id?: Id): IQuestion {
-    return {
-      id: modelData.id(id),
-      text: modelData.notes(), // TODO This may not be the correct type
-      data: modelData.notes(), // TODO This may not be the correct type
-      creatorId: modelData.id(),
-      updaterId: modelData.id(),
-      createdAt: modelData.timestamp(),
-      updatedAt: modelData.timestamp(),
-    };
-  }
 
   constructor(question: IQuestion, injector?: Injector) {
     super(question, injector);
