+++ src/app/models/Response.ts
@@ -1,7 +1,6 @@
 import { Injector } from "@angular/core";
 import { DATASET_ITEM, QUESTION, STUDY } from "@baw-api/ServiceTokens";
 import { DateTimeTimezone, Id } from "@interfaces/apiInterfaces";
-import { modelData } from "@test/helpers/faker";
 import { AbstractModel } from "./AbstractModel";
 import { Creator, HasOne } from "./AssociationDecorators";
 import { BawDateTime, BawPersistAttr } from "./AttributeDecorators";
@@ -45,18 +44,6 @@ export class Response extends AbstractModel implements IResponse {
   public question?: Question;
   @HasOne<Response>(STUDY, "studyId")
   public study?: Study;
-
-  public static generate(id?: Id): IResponse {
-    return {
-      id: modelData.id(id),
-      data: modelData.notes(),
-      datasetItemId: modelData.id(),
-      questionId: modelData.id(),
-      studyId: modelData.id(),
-      creatorId: modelData.id(),
-      createdAt: modelData.timestamp(),
-    };
-  }
 
   constructor(question: IResponse, injector?: Injector) {
     super(question, injector);
