+++ src/app/test/fakes/Response.ts
@@ -1,0 +1,15 @@
+import { Id } from "@interfaces/apiInterfaces";
+import { IResponse } from "@models/Response";
+import { modelData } from "@test/helpers/faker";
+
+export function generateResponse(id?: Id): IResponse {
+  return {
+    id: modelData.id(id),
+    data: modelData.notes(),
+    datasetItemId: modelData.id(),
+    questionId: modelData.id(),
+    studyId: modelData.id(),
+    creatorId: modelData.id(),
+    createdAt: modelData.timestamp(),
+  };
+}
