+++ src/app/test/fakes/TagGroup.ts
@@ -1,0 +1,13 @@
+import { Id } from "@interfaces/apiInterfaces";
+import { ITagGroup } from "@models/TagGroup";
+import { modelData } from "@test/helpers/faker";
+
+export function generateTagGroup(id?: Id): ITagGroup {
+  return {
+    id: modelData.id(id),
+    groupIdentifier: modelData.random.uuid(), // TODO Check
+    tagId: modelData.id(),
+    creatorId: modelData.id(),
+    createdAt: modelData.timestamp(),
+  };
+}
