+++ src/app/test/fakes/Tagging.ts
@@ -1,0 +1,15 @@
+import { Id } from "@interfaces/apiInterfaces";
+import { ITagging } from "@models/Tagging";
+import { modelData } from "@test/helpers/faker";
+
+export function generateTagging(id?: Id): ITagging {
+  return {
+    id: modelData.id(id),
+    audioEventId: modelData.id(),
+    tagId: modelData.id(),
+    creatorId: modelData.id(),
+    updaterId: modelData.id(),
+    createdAt: modelData.timestamp(),
+    updatedAt: modelData.timestamp(),
+  };
+}
